♻️ Reorganize note feature layout
