+++ packages/server/src/features/note/graphql/index.ts
@@ -1,0 +1,2 @@
+export { createAllNotesQueryResolver, noteResolvers } from './note.resolver.js';
+export { noteTypeDefs } from './note.type-defs.js';
