+++ packages/server/src/features/note/services/search.ts
