+++ packages/server/src/routes/mcp.ts
@@ -1,9 +1,13 @@
 import { Router } from 'express';
+import {
+    createMcpCreateNoteHandler,
+    createMcpDeleteNoteHandler,
+    createMcpUpdateNoteHandler,
+} from '../features/note/http/mcp.js';
 import type { AuthConfig } from '../modules/auth-mode.js';
 import type { McpAdminService } from '../modules/mcp-admin.js';
 import { createMcpAuthMiddleware } from '../modules/mcp-auth.js';
 import useAsync from '../modules/use-async.js';
-import { createMcpCreateNoteHandler, createMcpDeleteNoteHandler, createMcpUpdateNoteHandler } from '../views/note.js';
 import { createMcpCreateTagHandler } from '../views/tag.js';
 
 type McpRouteService = Pick<McpAdminService, 'getStatus' | 'validatePresentedToken'>;
