+++ packages/server/src/schema/index.ts
@@ -1,7 +1,7 @@
 import { makeExecutableSchema } from '@graphql-tools/schema';
+import { noteResolvers, noteTypeDefs } from '../features/note/graphql/index.js';
 import { cacheResolvers, cacheTypeDefs } from './cache/index.js';
 import { imageResolvers, imageTypeDefs } from './image/index.js';
-import { noteResolvers, noteTypeDefs } from './note/index.js';
 import { placeholderResolvers, placeholderTypeDefs } from './placeholder/index.js';
 import { reminderResolvers, reminderTypeDefs } from './reminder/index.js';
 import { tagResolvers, tagTypeDefs } from './tag/index.js';
