+++ packages/server/src/views/index.ts
@@ -1,5 +1,4 @@
 export * from './auth.js';
 export * from './image.js';
 export * from './mcp-admin.js';
-export * from './note.js';
 export * from './tag.js';
